% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{Age in years of individuals}
\usage{
age(x, reference = Sys.Date())
}
\arguments{
\item{x}{date(s), will be coerced with \code{\link{as.POSIXlt}}}

\item{reference}{reference date(s) (defaults to today), will be coerced with \code{\link{as.POSIXlt}}}
}
\value{
Integer (no decimals)
}
\description{
Calculates age in years based on a reference date, which is the sytem time at default.
}
\seealso{
\code{\link{age_groups}} to splits age into groups
}
