% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{antivirals}
\alias{antivirals}
\title{Data set with ~100 antivirals}
\format{A \code{\link{data.frame}} with 102 observations and 7 variables:
\describe{
  \item{\code{atc}}{ATC code (Anatomical Therapeutic Chemical) as defined by the WHOCC}
  \item{\code{name}}{Official name as used by WHONET/EARS-Net or the WHO}
  \item{\code{atc_group}}{Official pharmacological subgroup (3rd level ATC code) as defined by the WHOCC}
  \item{\code{oral_ddd}}{Defined Daily Dose (DDD), oral treatment}
  \item{\code{oral_units}}{Units of \code{oral_ddd}}
  \item{\code{iv_ddd}}{Defined Daily Dose (DDD), parenteral treatment}
  \item{\code{iv_units}}{Units of \code{iv_ddd}}
}}
\source{
World Health Organization (WHO) Collaborating Centre for Drug Statistics Methodology (WHOCC): \url{https://www.whocc.no/atc_ddd_index/}
}
\usage{
antivirals
}
\description{
A data set containing all antivirals, according to the ATC code group 'J05' (Antivirals for systemic use).
}
\section{WHOCC}{

\if{html}{\figure{logo_who.png}{options: height=60px style=margin-bottom:5px} \cr}
This package contains \strong{all ~550 antibiotic, antimycotic and antiviral drugs} and their Anatomical Therapeutic Chemical (ATC) codes, ATC groups and Defined Daily Dose (DDD) from the World Health Organization Collaborating Centre for Drug Statistics Methodology (WHOCC, \url{https://www.whocc.no}) and the Pharmaceuticals Community Register of the European Commission (\url{http://ec.europa.eu/health/documents/community-register/html/atc.htm}). 

These have become the gold standard for international drug utilisation monitoring and research.

The WHOCC is located in Oslo at the Norwegian Institute of Public Health and funded by the Norwegian government. The European Commission is the executive of the European Union and promotes its general interest.

\strong{NOTE: The WHOCC copyright does not allow use for commercial purposes, unlike any other info from this package. See \url{https://www.whocc.no/copyright_disclaimer/}.}
}

\section{Read more on our website!}{

On our website \url{https://msberends.gitlab.io/AMR} you can find \href{https://msberends.gitlab.io/AMR/articles/AMR.html}{a tutorial} about how to conduct AMR analysis, the \href{https://msberends.gitlab.io/AMR/reference}{complete documentation of all functions} (which reads a lot easier than here in R) and \href{https://msberends.gitlab.io/AMR/articles/WHONET.html}{an example analysis using WHONET data}.
}

\seealso{
\code{\link{antibiotics}} \code{\link{microorganisms}}
}
\keyword{datasets}
