% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended.R
\name{extended-functions}
\alias{extended-functions}
\alias{scale_type.mo}
\alias{scale_type.ab}
\title{Extended functions}
\usage{
scale_type.mo(x)

scale_type.ab(x)
}
\description{
These functions are extensions of functions in other packages.
}
\section{Read more on our website!}{

On our website \url{https://msberends.gitlab.io/AMR} you can find \href{https://msberends.gitlab.io/AMR/articles/AMR.html}{a tutorial} about how to conduct AMR analysis, the \href{https://msberends.gitlab.io/AMR/reference}{complete documentation of all functions} (which reads a lot easier than here in R) and \href{https://msberends.gitlab.io/AMR/articles/WHONET.html}{an example analysis using WHONET data}.
}

\keyword{internal}
