% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_ab.R
\name{guess_ab}
\alias{guess_ab}
\title{Guess antibiotic column}
\usage{
guess_ab(tbl = NULL, col = NULL, verbose = FALSE)
}
\arguments{
\item{tbl}{a \code{data.frame}}

\item{col}{a character to look for}

\item{verbose}{a logical to indicate whether additional info should be printed}
}
\description{
This tries to find a column name in a data set based on information from the \code{\link{antibiotics}} data set.
}
\section{Read more on our website!}{

\if{html}{\figure{logo.png}{options: height=40px style=margin-bottom:5px} \cr}
On our website \url{https://msberends.gitlab.io/AMR} you can find \href{https://msberends.gitlab.io/AMR/articles/AMR.html}{a omprehensive tutorial} about how to conduct AMR analysis and find \href{https://msberends.gitlab.io/AMR/reference}{the complete documentation of all functions}, which reads a lot easier than in R.
}

