% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R
\name{kurtosis}
\alias{kurtosis}
\alias{kurtosis.default}
\alias{kurtosis.matrix}
\alias{kurtosis.data.frame}
\title{Kurtosis of the sample}
\usage{
kurtosis(x, na.rm = FALSE)

\method{kurtosis}{default}(x, na.rm = FALSE)

\method{kurtosis}{matrix}(x, na.rm = FALSE)

\method{kurtosis}{data.frame}(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of values, a \code{matrix} or a \code{data frame}}

\item{na.rm}{a logical value indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\description{
Kurtosis is a measure of the "tailedness" of the probability distribution of a real-valued random variable.
}
\seealso{
\code{\link{skewness}}
}
