% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.R
\name{mo_failures}
\alias{mo_failures}
\title{Vector of failed coercion attempts}
\usage{
mo_failures()
}
\description{
Returns a vector of all failed attempts to coerce values to a valid MO code with \code{\link{as.mo}}.
}
\seealso{
\code{\link{as.mo}}
}
