% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo.R
\name{mo_renamed}
\alias{mo_renamed}
\title{Vector of taxonomic renamed items}
\usage{
mo_renamed()
}
\description{
Returns a vector of all renamed items of the last coercion to valid MO codes with \code{\link{as.mo}}.
}
\seealso{
\code{\link{as.mo}}
}
