% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trends}
\alias{trends}
\title{Detect trends using Machine Learning}
\usage{
trends(data, threshold_unique = 30, na.rm = TRUE, info = TRUE)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{threshold_unique}{do not analyse more unique \code{threshold_unique} items per variable}

\item{na.rm}{a logical value indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{info}{print relevant combinations to console}
}
\value{
A \code{list} with class \code{"trends"}
}
\description{
Test text
}
