% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_from_text.R
\name{ab_from_text}
\alias{ab_from_text}
\title{Retrieve antimicrobial drugs from text}
\usage{
ab_from_text(text, collapse = NULL, translate_ab = "name", ...)
}
\arguments{
\item{text}{text to analyse}

\item{collapse}{character to pass on to \code{paste(..., collapse = ...)} to only return one character per element of \code{text}, see Examples}

\item{translate_ab}{a column name of the \link{antibiotics} data set to translate the antibiotic abbreviations to, using \code{\link[=ab_property]{ab_property()}}. Defaults to "name", which is equal to using \code{TRUE}. Use a value \code{FALSE}, \code{NULL} or \code{NA} to prevent translation of the \verb{<ab>} code.}

\item{...}{parameters passed on to \code{\link[=as.ab]{as.ab()}}}
}
\description{
Use this function on e.g. clinical texts from health care records. It returns a vector of antimicrobial drugs found in the texts.
}
\details{
To use this for creating a new variable in a data set (e.g. with \code{mutate()}), it could be convenient to paste the outcome together with the \code{collapse} parameter so every value in your new variable will be a character of length 1:\cr
\code{df \%>\% mutate(abx = ab_from_text(clinical_text, collapse = "|"))}

This function is also internally used by \code{\link[=as.ab]{as.ab()}}, although it then only returns the first hit.
}
\examples{
# mind the bad spelling of amoxicillin in this line, 
# straight from a true health care record:
ab_from_text("28/03/2020 regular amoxicilliin 500mg po tds")

ab_from_text("administered amoxi/clav and cipro")
ab_from_text("administered amoxi/clav and cipro", collapse = ", ")

# if you want to know which antibiotic groups were administered, check it:
abx <- ab_from_text("administered amoxi/clav and cipro")
ab_group(abx)
}
