% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{Age in years of individuals}
\usage{
age(x, y = Sys.Date())
}
\arguments{
\item{x}{date(s) - will be coerced with \code{\link{as.POSIXlt}}}

\item{y}{reference date(s) - defaults to \code{\link{Sys.Date}} - will be coerced with \code{\link{as.POSIXlt}}}
}
\value{
Integer (no decimals)
}
\description{
Calculates age in years based on a reference date, which is the sytem time at default.
}
\seealso{
age_groups
}
