% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{Age in years of individuals}
\usage{
age(x, reference = Sys.Date())
}
\arguments{
\item{x}{date(s), will be coerced with \code{\link{as.POSIXlt}}}

\item{reference}{reference date(s) (defaults to today), will be coerced with \code{\link{as.POSIXlt}}}
}
\value{
Integer (no decimals)
}
\description{
Calculates age in years based on a reference date, which is the sytem time at default.
}
\section{Read more on our website!}{

\if{html}{\figure{logo.png}{options: height=40px style=margin-bottom:5px} \cr}
On our website \url{https://msberends.gitlab.io/AMR} you can find \href{https://msberends.gitlab.io/AMR/articles/AMR.html}{a omprehensive tutorial} about how to conduct AMR analysis and find \href{https://msberends.gitlab.io/AMR/reference}{the complete documentation of all functions}, which reads a lot easier than in R.
}

\seealso{
\code{\link{age_groups}} to splits age into groups
}
