% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_deprecated.R
\name{AMR-deprecated}
\alias{AMR-deprecated}
\alias{as.rsi}
\alias{is.rsi.eligible}
\title{Deprecated Functions}
\usage{
as.rsi(...)

is.rsi.eligible(...)
}
\description{
These functions are so-called '\link{Deprecated}'. \strong{They will be removed in a future release.} Using the functions will give a warning with the name of the function it has been replaced by (if there is one).
}
\keyword{internal}
