% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_deprecated.R
\docType{data}
\name{AMR-deprecated}
\alias{AMR-deprecated}
\alias{antibiotics}
\alias{ab_class}
\alias{ab_selector}
\title{Deprecated Functions, Arguments, or Datasets}
\format{
An object of class \code{deprecated_amr_dataset} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 497 rows and 14 columns.
}
\usage{
antibiotics

ab_class(...)

ab_selector(...)
}
\description{
These objects are so-called '\link{Deprecated}'. \strong{They will be removed in a future version of this package.} Using these will give a warning with the name of the alternative object it has been replaced by (if there is one).
}
\keyword{datasets}
\keyword{internal}
