% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{AMR-deprecated}
\alias{AMR-deprecated}
\alias{ratio}
\alias{abname}
\alias{atc_property}
\alias{atc_official}
\alias{ab_official}
\alias{atc_name}
\alias{atc_trivial_nl}
\alias{atc_tradenames}
\title{Deprecated functions}
\usage{
ratio(x, ratio)

abname(...)

atc_property(...)

atc_official(...)

ab_official(...)

atc_name(...)

atc_trivial_nl(...)

atc_tradenames(...)
}
\description{
These functions are so-called '\link{Deprecated}'. They will be removed in a future release. Using the functions will give a warning with the name of the function it has been replaced by (if there is one).
}
\section{Read more on our website!}{

On our website \url{https://msberends.gitlab.io/AMR} you can find \href{https://msberends.gitlab.io/AMR/articles/AMR.html}{a comprehensive tutorial} about how to conduct AMR analysis, the \href{https://msberends.gitlab.io/AMR/reference}{complete documentation of all functions} (which reads a lot easier than here in R) and \href{https://msberends.gitlab.io/AMR/articles/WHONET.html}{an example analysis using WHONET data}.
}

\keyword{internal}
