% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{AMR-deprecated}
\alias{AMR-deprecated}
\alias{p_symbol}
\alias{filter_first_weighted_isolate}
\alias{key_antibiotics}
\alias{key_antibiotics_equal}
\alias{filter_ab_class}
\alias{filter_aminoglycosides}
\alias{filter_betalactams}
\alias{filter_carbapenems}
\alias{filter_cephalosporins}
\alias{filter_1st_cephalosporins}
\alias{filter_2nd_cephalosporins}
\alias{filter_3rd_cephalosporins}
\alias{filter_4th_cephalosporins}
\alias{filter_5th_cephalosporins}
\alias{filter_fluoroquinolones}
\alias{filter_glycopeptides}
\alias{filter_macrolides}
\alias{filter_oxazolidinones}
\alias{filter_penicillins}
\alias{filter_tetracyclines}
\title{Deprecated Functions}
\usage{
p_symbol(p, emptychar = " ")

filter_first_weighted_isolate(
  x = NULL,
  col_date = NULL,
  col_patient_id = NULL,
  col_mo = NULL,
  ...
)

key_antibiotics(
  x = NULL,
  col_mo = NULL,
  universal_1 = guess_ab_col(x, "amoxicillin"),
  universal_2 = guess_ab_col(x, "amoxicillin/clavulanic acid"),
  universal_3 = guess_ab_col(x, "cefuroxime"),
  universal_4 = guess_ab_col(x, "piperacillin/tazobactam"),
  universal_5 = guess_ab_col(x, "ciprofloxacin"),
  universal_6 = guess_ab_col(x, "trimethoprim/sulfamethoxazole"),
  GramPos_1 = guess_ab_col(x, "vancomycin"),
  GramPos_2 = guess_ab_col(x, "teicoplanin"),
  GramPos_3 = guess_ab_col(x, "tetracycline"),
  GramPos_4 = guess_ab_col(x, "erythromycin"),
  GramPos_5 = guess_ab_col(x, "oxacillin"),
  GramPos_6 = guess_ab_col(x, "rifampin"),
  GramNeg_1 = guess_ab_col(x, "gentamicin"),
  GramNeg_2 = guess_ab_col(x, "tobramycin"),
  GramNeg_3 = guess_ab_col(x, "colistin"),
  GramNeg_4 = guess_ab_col(x, "cefotaxime"),
  GramNeg_5 = guess_ab_col(x, "ceftazidime"),
  GramNeg_6 = guess_ab_col(x, "meropenem"),
  warnings = TRUE,
  ...
)

key_antibiotics_equal(
  y,
  z,
  type = "keyantimicrobials",
  ignore_I = TRUE,
  points_threshold = 2,
  info = FALSE,
  na.rm = TRUE,
  ...
)

filter_ab_class(
  x,
  ab_class,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_aminoglycosides(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_betalactams(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_carbapenems(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_cephalosporins(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_1st_cephalosporins(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_2nd_cephalosporins(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_3rd_cephalosporins(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_4th_cephalosporins(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_5th_cephalosporins(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_fluoroquinolones(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_glycopeptides(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_macrolides(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_oxazolidinones(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_penicillins(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)

filter_tetracyclines(
  x,
  result = NULL,
  scope = "any",
  only_rsi_columns = FALSE,
  ...
)
}
\description{
These functions are so-called '\link{Deprecated}'. \strong{They will be removed in a future release.} Using the functions will give a warning with the name of the function it has been replaced by (if there is one).
}
\details{
All antibiotic class selectors (such as \code{\link[=carbapenems]{carbapenems()}}, \code{\link[=aminoglycosides]{aminoglycosides()}}) can now be used for filtering as well, making all their accompanying \verb{filter_*()} functions redundant (such as \code{\link[=filter_carbapenems]{filter_carbapenems()}}, \code{\link[=filter_aminoglycosides]{filter_aminoglycosides()}}).
}
\section{Retired Lifecycle}{

\if{html}{\figure{lifecycle_retired.svg}{options: style=margin-bottom:5px} \cr}
The \link[=lifecycle]{lifecycle} of this function is \strong{retired}. A retired function is no longer under active development, and (if appropiate) a better alternative is available. No new arguments will be added, and only the most critical bugs will be fixed. In a future version, this function will be removed.
}

\section{Read more on Our Website!}{

On our website \url{https://msberends.github.io/AMR/} you can find \href{https://msberends.github.io/AMR/articles/AMR.html}{a comprehensive tutorial} about how to conduct AMR data analysis, the \href{https://msberends.github.io/AMR/reference/}{complete documentation of all functions} and \href{https://msberends.github.io/AMR/articles/WHONET.html}{an example analysis using WHONET data}.
}

\keyword{internal}
