% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_deprecated.R
\docType{data}
\name{AMR-deprecated}
\alias{AMR-deprecated}
\alias{NA_rsi_}
\alias{as.rsi}
\alias{facet_rsi}
\alias{geom_rsi}
\alias{ggplot_rsi}
\alias{ggplot_rsi_predict}
\alias{is.rsi}
\alias{is.rsi.eligible}
\alias{labels_rsi_count}
\alias{n_rsi}
\alias{random_rsi}
\alias{rsi_df}
\alias{rsi_predict}
\alias{scale_rsi_colours}
\alias{theme_rsi}
\title{Deprecated Functions}
\format{
An object of class \code{rsi} (inherits from \code{ordered}, \code{factor}) of length 1.
}
\usage{
NA_rsi_

as.rsi(x, ...)

facet_rsi(...)

geom_rsi(...)

ggplot_rsi(...)

ggplot_rsi_predict(...)

is.rsi(...)

is.rsi.eligible(...)

labels_rsi_count(...)

n_rsi(...)

random_rsi(...)

rsi_df(...)

rsi_predict(...)

scale_rsi_colours(...)

theme_rsi(...)
}
\description{
These functions are so-called '\link{Deprecated}'. \strong{They will be removed in a future release.} Using the functions will give a warning with the name of the function it has been replaced by (if there is one).
}
\keyword{internal}
