% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availability.R
\name{availability}
\alias{availability}
\title{Check availability of columns}
\usage{
availability(tbl)
}
\arguments{
\item{tbl}{a \code{data.frame} or \code{list}}
}
\value{
\code{data.frame} with column names of \code{tbl} as row names and columns: \code{percent_IR}, \code{count}, \code{percent}, \code{visual_availability}.
}
\description{
Easy check for availability of columns in a data set. This makes it easy to get an idea of which antibiotic combination can be used for calculation with e.g. \code{\link{portion_IR}}.
}
\examples{
availability(septic_patients)

library(dplyr)
septic_patients \%>\% availability()

septic_patients \%>\%
  select_if(is.rsi) \%>\%
  availability()

septic_patients \%>\%
  filter(mo == as.mo("E. coli")) \%>\%
  select_if(is.rsi) \%>\%
  availability()
}
