% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue_of_life.R
\name{catalogue_of_life_version}
\alias{catalogue_of_life_version}
\title{Version info of included Catalogue of Life}
\usage{
catalogue_of_life_version()
}
\value{
a \code{\link{list}}, which prints in pretty format
}
\description{
This function returns information about the included data from the Catalogue of Life.
}
\details{
For DSMZ, see \link{microorganisms}.
}
\section{Catalogue of Life}{

\if{html}{\figure{logo_col.png}{options: height=40px style=margin-bottom:5px} \cr}
This package contains the complete taxonomic tree of almost all microorganisms (~70,000 species) from the authoritative and comprehensive Catalogue of Life (\url{http://www.catalogueoflife.org}). The Catalogue of Life is the most comprehensive and authoritative global index of species currently available.

\link[=catalogue_of_life]{Click here} for more information about the included taxa. Check which version of the Catalogue of Life was included in this package with \code{\link[=catalogue_of_life_version]{catalogue_of_life_version()}}.
}

\section{Read more on our website!}{

On our website \url{https://msberends.github.io/AMR} you can find \href{https://msberends.github.io/AMR/articles/AMR.html}{a comprehensive tutorial} about how to conduct AMR analysis, the \href{https://msberends.github.io/AMR/reference}{complete documentation of all functions} (which reads a lot easier than here in R) and \href{https://msberends.github.io/AMR/articles/WHONET.html}{an example analysis using WHONET data}.
}

\seealso{
\link{microorganisms}
}
