% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join}
\alias{join}
\alias{inner_join_bactlist}
\alias{inner_join}
\alias{left_join_bactlist}
\alias{right_join_bactlist}
\alias{full_join_bactlist}
\alias{semi_join_bactlist}
\alias{anti_join_bactlist}
\title{Join van tabel en \code{bactlist}}
\usage{
inner_join_bactlist(x, by = "bactid", ...)

left_join_bactlist(x, by = "bacteriecode", ...)

right_join_bactlist(x, by = "bacteriecode", ...)

full_join_bactlist(x, by = "bacteriecode", ...)

semi_join_bactlist(x, by = "bacteriecode", ...)

anti_join_bactlist(x, by = "bacteriecode", ...)
}
\arguments{
\item{x}{existing table to join}

\item{by}{a variable to join by - could be a column name of \code{x} with values that exist in \code{bactlist$bactid} (like \code{by = "bacteria_id"}), or another column in \code{\link{bactlist}} (but then it should be named, like \code{by = c("my_genus_species" = "fullname")})}

\item{...}{other parameters to pass trhough to \code{dplyr::\link[dplyr]{join}}.}
}
\description{
Join the list of microorganisms \code{\link{bactlist}} easily to an existing table.
}
\details{
As opposed to the \code{\link[dplyr]{join}} functions of \code{dplyr}, at default existing columns will get a suffix \code{"2"} and the newly joined columns will not get a suffix. See \code{\link[dplyr]{join}} for more information.
}
