% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R
\name{kurtosis}
\alias{kurtosis}
\alias{kurtosis.default}
\alias{kurtosis.matrix}
\alias{kurtosis.data.frame}
\title{Kurtosis of the sample}
\usage{
kurtosis(x, na.rm = FALSE)

\method{kurtosis}{default}(x, na.rm = FALSE)

\method{kurtosis}{matrix}(x, na.rm = FALSE)

\method{kurtosis}{data.frame}(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of values, a \code{\link{matrix}} or a \code{\link{data.frame}}}

\item{na.rm}{a logical value indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\description{
Kurtosis is a measure of the "tailedness" of the probability distribution of a real-valued random variable.
}
\section{Questioning lifecycle}{

\if{html}{\figure{lifecycle_questioning.svg}{options: style=margin-bottom:5px} \cr}
The \link[=lifecycle]{lifecycle} of this function is \strong{questioning}. This function might be no longer be optimal approach, or is it questionable whether this function should be in this \code{AMR} package at all.
}

\section{Read more on our website!}{

On our website \url{https://msberends.github.io/AMR} you can find \href{https://msberends.github.io/AMR/articles/AMR.html}{a comprehensive tutorial} about how to conduct AMR analysis, the \href{https://msberends.github.io/AMR/reference}{complete documentation of all functions} (which reads a lot easier than here in R) and \href{https://msberends.github.io/AMR/articles/WHONET.html}{an example analysis using WHONET data}.
}

\seealso{
\code{\link[=skewness]{skewness()}}
}
