% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdro.R
\name{MDRO}
\alias{MDRO}
\alias{BRMO}
\alias{MRGN}
\title{Determine multidrug-resistant organisms (MDRO)}
\usage{
MDRO(tbl, country, col_bactid = "bactid", info = TRUE,
  aminoglycosides = c("gent", "tobr", "kana"), quinolones = c("cipr",
  "norf"), carbapenems = c("imip", "mero", "erta"), ceftazidime = "cfta",
  piperacillin = "pita", trimethoprim_sulfa = "trsu", penicillin = "peni",
  vancomycin = "vanc")

BRMO(tbl, country = "nl", ...)

MRGN(tbl, country = "de", ...)
}
\arguments{
\item{tbl}{table with antibiotic columns, like e.g. \code{amox} and \code{amcl}}

\item{country}{country to determine guidelines. Should be a code from the \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2#Officially_assigned_code_elements}{list of ISO 3166-1 alpha-2 country codes}. Case-insensitive. Currently supported are \code{de} (Germany) and \code{nl} (the Netherlands).}

\item{col_bactid}{column name of the bacteria ID in \code{tbl} - values of this column should be present in \code{microorganisms$bactid}, see \code{\link{microorganisms}}}

\item{info}{print progress}

\item{aminoglycosides, quinolones, carbapenems}{character vector with column names of antibiotics}

\item{ceftazidime, piperacillin, trimethoprim_sulfa, penicillin, vancomycin}{column names of antibiotics}

\item{...}{parameters that are passed on to \code{MDR}}
}
\value{
Ordered factor with values \code{Positive}, \code{Unconfirmed}, \code{Negative}.
}
\description{
Determine which isolates are multidrug-resistant organisms (MDRO) according to country-specific guidelines.
}
