% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdro.R
\name{mdro}
\alias{mdro}
\alias{brmo}
\alias{mrgn}
\alias{eucast_exceptional_phenotypes}
\title{Determine multidrug-resistant organisms (MDRO)}
\usage{
mdro(tbl, country = NULL, col_mo = NULL, info = TRUE,
  amcl = guess_ab(), amik = guess_ab(), amox = guess_ab(),
  ampi = guess_ab(), azit = guess_ab(), aztr = guess_ab(),
  cefa = guess_ab(), cfra = guess_ab(), cfep = guess_ab(),
  cfot = guess_ab(), cfox = guess_ab(), cfta = guess_ab(),
  cftr = guess_ab(), cfur = guess_ab(), chlo = guess_ab(),
  cipr = guess_ab(), clar = guess_ab(), clin = guess_ab(),
  clox = guess_ab(), coli = guess_ab(), czol = guess_ab(),
  dapt = guess_ab(), doxy = guess_ab(), erta = guess_ab(),
  eryt = guess_ab(), fosf = guess_ab(), fusi = guess_ab(),
  gent = guess_ab(), imip = guess_ab(), kana = guess_ab(),
  levo = guess_ab(), linc = guess_ab(), line = guess_ab(),
  mero = guess_ab(), metr = guess_ab(), mino = guess_ab(),
  moxi = guess_ab(), nali = guess_ab(), neom = guess_ab(),
  neti = guess_ab(), nitr = guess_ab(), novo = guess_ab(),
  norf = guess_ab(), oflo = guess_ab(), peni = guess_ab(),
  pipe = guess_ab(), pita = guess_ab(), poly = guess_ab(),
  qida = guess_ab(), rifa = guess_ab(), roxi = guess_ab(),
  siso = guess_ab(), teic = guess_ab(), tetr = guess_ab(),
  tica = guess_ab(), tige = guess_ab(), tobr = guess_ab(),
  trim = guess_ab(), trsu = guess_ab(), vanc = guess_ab())

brmo(..., country = "nl")

mrgn(tbl, country = "de", ...)

eucast_exceptional_phenotypes(tbl, country = "EUCAST", ...)
}
\arguments{
\item{tbl}{table with antibiotic columns, like e.g. \code{amox} and \code{amcl}}

\item{country}{country code to determine guidelines. EUCAST rules will be used when left empty, see Details. Should be or a code from the \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2#Officially_assigned_code_elements}{list of ISO 3166-1 alpha-2 country codes}. Case-insensitive. Currently supported are \code{de} (Germany) and \code{nl} (the Netherlands).}

\item{col_mo}{column name of the unique IDs of the microorganisms (see \code{\link{mo}}), defaults to the first column of class \code{mo}. Values will be coerced using \code{\link{as.mo}}.}

\item{info}{print progress}

\item{amcl}{column name of an antibiotic, see Antibiotics}

\item{amik}{column name of an antibiotic, see Antibiotics}

\item{amox}{column name of an antibiotic, see Antibiotics}

\item{ampi}{column name of an antibiotic, see Antibiotics}

\item{azit}{column name of an antibiotic, see Antibiotics}

\item{aztr}{column name of an antibiotic, see Antibiotics}

\item{cefa}{column name of an antibiotic, see Antibiotics}

\item{cfra}{column name of an antibiotic, see Antibiotics}

\item{cfep}{column name of an antibiotic, see Antibiotics}

\item{cfot}{column name of an antibiotic, see Antibiotics}

\item{cfox}{column name of an antibiotic, see Antibiotics}

\item{cfta}{column name of an antibiotic, see Antibiotics}

\item{cftr}{column name of an antibiotic, see Antibiotics}

\item{cfur}{column name of an antibiotic, see Antibiotics}

\item{chlo}{column name of an antibiotic, see Antibiotics}

\item{cipr}{column name of an antibiotic, see Antibiotics}

\item{clar}{column name of an antibiotic, see Antibiotics}

\item{clin}{column name of an antibiotic, see Antibiotics}

\item{clox}{column name of an antibiotic, see Antibiotics}

\item{coli}{column name of an antibiotic, see Antibiotics}

\item{czol}{column name of an antibiotic, see Antibiotics}

\item{dapt}{column name of an antibiotic, see Antibiotics}

\item{doxy}{column name of an antibiotic, see Antibiotics}

\item{erta}{column name of an antibiotic, see Antibiotics}

\item{eryt}{column name of an antibiotic, see Antibiotics}

\item{fosf}{column name of an antibiotic, see Antibiotics}

\item{fusi}{column name of an antibiotic, see Antibiotics}

\item{gent}{column name of an antibiotic, see Antibiotics}

\item{imip}{column name of an antibiotic, see Antibiotics}

\item{kana}{column name of an antibiotic, see Antibiotics}

\item{levo}{column name of an antibiotic, see Antibiotics}

\item{linc}{column name of an antibiotic, see Antibiotics}

\item{line}{column name of an antibiotic, see Antibiotics}

\item{mero}{column name of an antibiotic, see Antibiotics}

\item{metr}{column name of an antibiotic, see Antibiotics}

\item{mino}{column name of an antibiotic, see Antibiotics}

\item{moxi}{column name of an antibiotic, see Antibiotics}

\item{nali}{column name of an antibiotic, see Antibiotics}

\item{neom}{column name of an antibiotic, see Antibiotics}

\item{neti}{column name of an antibiotic, see Antibiotics}

\item{nitr}{column name of an antibiotic, see Antibiotics}

\item{novo}{column name of an antibiotic, see Antibiotics}

\item{norf}{column name of an antibiotic, see Antibiotics}

\item{oflo}{column name of an antibiotic, see Antibiotics}

\item{peni}{column name of an antibiotic, see Antibiotics}

\item{pipe}{column name of an antibiotic, see Antibiotics}

\item{pita}{column name of an antibiotic, see Antibiotics}

\item{poly}{column name of an antibiotic, see Antibiotics}

\item{qida}{column name of an antibiotic, see Antibiotics}

\item{rifa}{column name of an antibiotic, see Antibiotics}

\item{roxi}{column name of an antibiotic, see Antibiotics}

\item{siso}{column name of an antibiotic, see Antibiotics}

\item{teic}{column name of an antibiotic, see Antibiotics}

\item{tetr}{column name of an antibiotic, see Antibiotics}

\item{tica}{column name of an antibiotic, see Antibiotics}

\item{tige}{column name of an antibiotic, see Antibiotics}

\item{tobr}{column name of an antibiotic, see Antibiotics}

\item{trim}{column name of an antibiotic, see Antibiotics}

\item{trsu}{column name of an antibiotic, see Antibiotics}

\item{vanc}{column name of an antibiotic, see Antibiotics}

\item{...}{parameters that are passed on to methods}
}
\value{
Ordered factor with levels \code{Negative < Positive, unconfirmed < Positive}.
}
\description{
Determine which isolates are multidrug-resistant organisms (MDRO) according to country-specific guidelines.
}
\details{
When \code{country} will be left blank, guidelines will be taken from EUCAST Expert Rules Version 3.1 "Intrinsic Resistance and Exceptional Phenotypes Tables" (\url{http://www.eucast.org/fileadmin/src/media/PDFs/EUCAST_files/Expert_Rules/Expert_rules_intrinsic_exceptional_V3.1.pdf}).
}
\section{Antibiotics}{

To define antibiotics column names, leave as it is to determine it automatically with \code{\link{guess_ab}} or input a text (case-insensitive) or use \code{NULL} to skip a column (e.g. \code{tica = NULL}). Non-existing columns will anyway be skipped with a warning.

Abbrevations of the column containing antibiotics in the form: \strong{abbreviation}: generic name (\emph{ATC code})

 \strong{amcl}: amoxicillin+clavulanic acid (\emph{J01CR02}),
 \strong{amik}: amikacin (\emph{J01GB06}),
 \strong{amox}: amoxicillin (\emph{J01CA04}),
 \strong{ampi}: ampicillin (\emph{J01CA01}),
 \strong{azit}: azithromycin (\emph{J01FA10}),
 \strong{azlo}: azlocillin (\emph{J01CA09}),
 \strong{aztr}: aztreonam (\emph{J01DF01}),
 \strong{cefa}: cefaloridine (\emph{J01DB02}),
 \strong{cfep}: cefepime (\emph{J01DE01}),
 \strong{cfot}: cefotaxime (\emph{J01DD01}),
 \strong{cfox}: cefoxitin (\emph{J01DC01}),
 \strong{cfra}: cefradine (\emph{J01DB09}),
 \strong{cfta}: ceftazidime (\emph{J01DD02}),
 \strong{cftr}: ceftriaxone (\emph{J01DD04}),
 \strong{cfur}: cefuroxime (\emph{J01DC02}),
 \strong{chlo}: chloramphenicol (\emph{J01BA01}),
 \strong{cipr}: ciprofloxacin (\emph{J01MA02}),
 \strong{clar}: clarithromycin (\emph{J01FA09}),
 \strong{clin}: clindamycin (\emph{J01FF01}),
 \strong{clox}: flucloxacillin (\emph{J01CF05}),
 \strong{coli}: colistin (\emph{J01XB01}),
 \strong{czol}: cefazolin (\emph{J01DB04}),
 \strong{dapt}: daptomycin (\emph{J01XX09}),
 \strong{doxy}: doxycycline (\emph{J01AA02}),
 \strong{erta}: ertapenem (\emph{J01DH03}),
 \strong{eryt}: erythromycin (\emph{J01FA01}),
 \strong{fosf}: fosfomycin (\emph{J01XX01}),
 \strong{fusi}: fusidic acid (\emph{J01XC01}),
 \strong{gent}: gentamicin (\emph{J01GB03}),
 \strong{imip}: imipenem (\emph{J01DH51}),
 \strong{kana}: kanamycin (\emph{J01GB04}),
 \strong{levo}: levofloxacin (\emph{J01MA12}),
 \strong{linc}: lincomycin (\emph{J01FF02}),
 \strong{line}: linezolid (\emph{J01XX08}),
 \strong{mero}: meropenem (\emph{J01DH02}),
 \strong{mezl}: mezlocillin (\emph{J01CA10}),
 \strong{mino}: minocycline (\emph{J01AA08}),
 \strong{moxi}: moxifloxacin (\emph{J01MA14}),
 \strong{nali}: nalidixic acid (\emph{J01MB02}),
 \strong{neom}: neomycin (\emph{J01GB05}),
 \strong{neti}: netilmicin (\emph{J01GB07}),
 \strong{nitr}: nitrofurantoin (\emph{J01XE01}),
 \strong{norf}: norfloxacin (\emph{J01MA06}),
 \strong{novo}: novobiocin (an ATCvet code: \emph{QJ01XX95}),
 \strong{oflo}: ofloxacin (\emph{J01MA01}),
 \strong{peni}: penicillin (\emph{J01RA01}),
 \strong{pipe}: piperacillin (\emph{J01CA12}),
 \strong{pita}: piperacillin+tazobactam (\emph{J01CR05}),
 \strong{poly}: polymyxin B (\emph{J01XB02}),
 \strong{pris}: pristinamycin (\emph{J01FG01}),
 \strong{qida}: quinupristin/dalfopristin (\emph{J01FG02}),
 \strong{rifa}: rifampicin (\emph{J04AB02}),
 \strong{roxi}: roxithromycin (\emph{J01FA06}),
 \strong{siso}: sisomicin (\emph{J01GB08}),
 \strong{teic}: teicoplanin (\emph{J01XA02}),
 \strong{tetr}: tetracycline (\emph{J01AA07}),
 \strong{tica}: ticarcillin (\emph{J01CA13}),
 \strong{tige}: tigecycline (\emph{J01AA12}),
 \strong{tobr}: tobramycin (\emph{J01GB01}),
 \strong{trim}: trimethoprim (\emph{J01EA01}),
 \strong{trsu}: sulfamethoxazole and trimethoprim (\emph{J01EE01}),
 \strong{vanc}: vancomycin (\emph{J01XA01}).
}

\section{Read more on our website!}{

\if{html}{\figure{logo.png}{options: height=40px style=margin-bottom:5px} \cr}
On our website \url{https://msberends.gitlab.io/AMR} you can find \href{https://msberends.gitlab.io/AMR/articles/AMR.html}{a omprehensive tutorial} about how to conduct AMR analysis and find \href{https://msberends.gitlab.io/AMR/reference}{the complete documentation of all functions}, which reads a lot easier than in R.
}

\examples{
library(dplyr)

septic_patients \%>\%
  mutate(EUCAST = mdro(.),
         BRMO = brmo(.))
}
