% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microorganisms.old}
\alias{microorganisms.old}
\title{Data set with old taxonomic data from ITIS}
\format{A \code{\link{data.frame}} with 58 observations and 5 variables:
\describe{
  \item{\code{tsn}}{Old Taxonomic Serial Number (TSN), as defined by ITIS}
  \item{\code{name}}{Old taxonomic name of the microorganism as found in ITIS, see Source}
  \item{\code{tsn_new}}{New Taxonomic Serial Number (TSN), as defined by ITIS}
  \item{\code{authors}}{Authors responsible for renaming as found in ITIS, see Source}
  \item{\code{year}}{Year in which the literature was published about the renaming as found in ITIS, see Source}
}}
\source{
[3] Integrated Taxonomic Information System (ITIS) on-line database, \url{https://www.itis.gov}.
}
\usage{
microorganisms.old
}
\description{
A data set containing old, previously valid, taxonomic names. This data set is used internally by \code{\link{as.mo}}.
}
\section{ITIS}{

\if{html}{\figure{itis_logo.jpg}{options: height=60px style=margin-bottom:5px} \cr}
This \code{AMR} package contains the \strong{complete microbial taxonomic data} from the publicly available Integrated Taxonomic Information System (ITIS, https://www.itis.gov). ITIS is a partnership of U.S., Canadian, and Mexican agencies and taxonomic specialists [3]. The complete taxonomic kingdoms Bacteria, Fungi and Protozoa (from subkingdom to the subspecies level) are included in this package.
}

\seealso{
\code{\link{as.mo}} \code{\link{mo_property}} \code{\link{microorganisms}}
}
\keyword{datasets}
