% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microorganisms}
\alias{microorganisms}
\title{Data set with ~60,000 microorganisms}
\format{A \code{\link{data.frame}} with 56,672 observations and 14 variables:
\describe{
  \item{\code{mo}}{ID of microorganism as used by this package}
  \item{\code{col_id}}{Catalogue of Life ID}
  \item{\code{fullname}}{Full name, like \code{"Echerichia coli"}}
  \item{\code{kingdom}}{Taxonomic kingdom of the microorganism}
  \item{\code{phylum}}{Taxonomic phylum of the microorganism}
  \item{\code{class}}{Taxonomic class of the microorganism}
  \item{\code{order}}{Taxonomic order of the microorganism}
  \item{\code{family}}{Taxonomic family of the microorganism}
  \item{\code{genus}}{Taxonomic genus of the microorganism}
  \item{\code{species}}{Taxonomic species of the microorganism}
  \item{\code{subspecies}}{Taxonomic subspecies of the microorganism}
  \item{\code{rank}}{Taxonomic rank of the microorganism, like \code{"species"} or \code{"genus"}}
  \item{\code{ref}}{Author(s) and year of concerning scientific publication}
  \item{\code{species_id}}{ID of the species as used by the Catalogue of Life}
}}
\source{
Catalogue of Life: Annual Checklist (public online database), \url{www.catalogueoflife.org}.
}
\usage{
microorganisms
}
\description{
A data set containing the microbial taxonomy of six kingdoms from the Catalogue of Life. MO codes can be looked up using \code{\link{as.mo}}.
}
\details{
Manually added were:
\itemize{
  \item{9 species of \emph{Streptococcus} (beta haemolytic groups A, B, C, D, F, G, H, K and unspecified)}
  \item{2 species of \emph{Staphylococcus} (coagulase-negative [CoNS] and coagulase-positive [CoPS])}
  \item{2 other undefined (unknown Gram negatives and unknown Gram positives)}
}
}
\section{Catalogue of Life}{

\if{html}{\figure{logo_col.png}{options: height=60px style=margin-bottom:5px} \cr}
This package contains the complete taxonomic tree of almost all microorganisms from the authoritative and comprehensive Catalogue of Life (\url{http://www.catalogueoflife.org}). This data is updated annually - check the included version with \code{\link{catalogue_of_life_version}}.

Included are:
\itemize{
  \item{All ~55,000 (sub)species from the kingdoms of Archaea, Bacteria, Protozoa and Viruses}
  \item{All ~3,000 (sub)species from these orders of the kingdom of Fungi: Eurotiales, Onygenales, Pneumocystales, Saccharomycetales and Schizosaccharomycetales. The kingdom of Fungi is a very large taxon with almost 300,000 different species, of which most are not microbial. Including everything tremendously slows down our algortihms, and not all fungi fit the scope of this package. By only including the aforementioned taxonomic orders, the most relevant species are covered (like genera \emph{Aspergillus}, \emph{Candida}, \emph{Pneumocystis}, \emph{Saccharomyces} and \emph{Trichophyton}).}
  \item{All ~15,000 previously accepted names of inckuded (sub)species that have been taxonomically renamed}
  \item{The complete taxonomic tree of all included (sub)species: from kingdom to subspecies}
  \item{The responsible author(s) and year of scientific publication}
}

The Catalogue of Life (\url{http://www.catalogueoflife.org}) is the most comprehensive and authoritative global index of species currently available. It holds essential information on the names, relationships and distributions of over 1.6 million species. The Catalogue of Life is used to support the major biodiversity and conservation information services such as the Global Biodiversity Information Facility (GBIF), Encyclopedia of Life (EoL) and the International Union for Conservation of Nature Red List. It is recognised by the Convention on Biological Diversity as a significant component of the Global Taxonomy Initiative and a contribution to Target 1 of the Global Strategy for Plant Conservation.

The syntax used to transform the original data to a cleansed R format, can be found here: \url{https://gitlab.com/msberends/AMR/blob/master/reproduction_of_microorganisms.R}.
}

\section{Read more on our website!}{

\if{html}{\figure{logo.png}{options: height=40px style=margin-bottom:5px} \cr}
On our website \url{https://msberends.gitlab.io/AMR} you can find \href{https://msberends.gitlab.io/AMR/articles/AMR.html}{a comprehensive tutorial} about how to conduct AMR analysis, the \href{https://msberends.gitlab.io/AMR/reference}{complete documentation of all functions} (which reads a lot easier than here in R) and \href{https://msberends.gitlab.io/AMR/articles/WHONET.html}{an example analysis using WHONET data}.
}

\seealso{
\code{\link{as.mo}}, \code{\link{mo_property}}, \code{\link{microorganisms.codes}}
}
\keyword{datasets}
