% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microorganisms}
\alias{microorganisms}
\title{Data set with ~60,000 microorganisms}
\format{A \code{\link{data.frame}} with 59,985 observations and 15 variables:
\describe{
  \item{\code{mo}}{ID of microorganism as used by this package}
  \item{\code{col_id}}{Catalogue of Life ID}
  \item{\code{fullname}}{Full name, like \code{"Echerichia coli"}}
  \item{\code{kingdom}}{Taxonomic kingdom of the microorganism}
  \item{\code{phylum}}{Taxonomic phylum of the microorganism}
  \item{\code{class}}{Taxonomic class of the microorganism}
  \item{\code{order}}{Taxonomic order of the microorganism}
  \item{\code{family}}{Taxonomic family of the microorganism}
  \item{\code{genus}}{Taxonomic genus of the microorganism}
  \item{\code{species}}{Taxonomic species of the microorganism}
  \item{\code{subspecies}}{Taxonomic subspecies of the microorganism}
  \item{\code{rank}}{Taxonomic rank of the microorganism, like \code{"species"} or \code{"genus"}}
  \item{\code{ref}}{Author(s) and year of concerning scientific publication}
  \item{\code{species_id}}{ID of the species as used by the Catalogue of Life}
  \item{\code{prevalence}}{Prevalence of the microorganism, see \code{?as.mo}}
}}
\source{
Catalogue of Life: Annual Checklist (public online database), \url{www.catalogueoflife.org}.
}
\usage{
microorganisms
}
\description{
A data set containing the microbial taxonomy of six kingdoms from the Catalogue of Life. MO codes can be looked up using \code{\link{as.mo}}.
}
\details{
Manually added were:
\itemize{
  \item{9 species of \emph{Streptococcus} (beta haemolytic groups A, B, C, D, F, G, H, K and unspecified)}
  \item{2 species of \emph{Staphylococcus} (coagulase-negative [CoNS] and coagulase-positive [CoPS])}
  \item{2 other undefined (unknown Gram negatives and unknown Gram positives)}
}
}
\section{Catalogue of Life}{

\if{html}{\figure{logo_col.png}{options: height=40px style=margin-bottom:5px} \cr}
This package contains the complete taxonomic tree of almost all microorganisms (~60,000 species) from the authoritative and comprehensive Catalogue of Life (\url{http://www.catalogueoflife.org}). The Catalogue of Life is the most comprehensive and authoritative global index of species currently available.

\link[=catalogue_of_life]{Click here} for more information about the included taxa. The Catalogue of Life releases updates annually; check which version was included in this package with \code{\link{catalogue_of_life_version}()}.
}

\section{Read more on our website!}{

On our website \url{https://msberends.gitlab.io/AMR} you can find \href{https://msberends.gitlab.io/AMR/articles/AMR.html}{a comprehensive tutorial} about how to conduct AMR analysis, the \href{https://msberends.gitlab.io/AMR/reference}{complete documentation of all functions} (which reads a lot easier than here in R) and \href{https://msberends.gitlab.io/AMR/articles/WHONET.html}{an example analysis using WHONET data}.
}

\seealso{
\code{\link{as.mo}}, \code{\link{mo_property}}, \code{\link{microorganisms.codes}}
}
\keyword{datasets}
