% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_property.R
\name{mo_property}
\alias{mo_property}
\alias{mo_family}
\alias{mo_genus}
\alias{mo_species}
\alias{mo_subspecies}
\alias{mo_fullname}
\alias{mo_type}
\alias{mo_gramstain}
\alias{mo_aerobic}
\alias{mo_type_nl}
\alias{mo_gramstain_nl}
\title{Property of a microorganism}
\usage{
mo_property(x, property = "fullname")

mo_family(x)

mo_genus(x)

mo_species(x)

mo_subspecies(x)

mo_fullname(x)

mo_type(x)

mo_gramstain(x)

mo_aerobic(x)

mo_type_nl(x)

mo_gramstain_nl(x)
}
\arguments{
\item{x}{a (vector of a) valid \code{\link{bactid}} or any text that can be coerced to a valid bactid with \code{\link{as.bactid}}}

\item{property}{one of the column names of one of the \code{\link{microorganisms}} data set, like \code{"bactid"}, \code{"bactsys"}, \code{"family"}, \code{"genus"}, \code{"species"}, \code{"fullname"}, \code{"gramstain"} and \code{"aerobic"}}
}
\description{
Use these functions to return a specific property of a microorganism from the \code{\link{microorganisms}} data set, based on their \code{bactid}. Get such an ID with \code{\link{as.bactid}}.
}
\examples{
# All properties
mo_family("E. coli")       # Enterobacteriaceae
mo_genus("E. coli")        # Escherichia
mo_species("E. coli")      # coli
mo_subspecies("E. coli")   # <NA>
mo_fullname("E. coli")     # Escherichia coli
mo_type("E. coli")         # Bacteria
mo_gramstain("E. coli")    # Negative rods
mo_aerobic("E. coli")      # TRUE
mo_type_nl("E. coli")      # Bacterie
mo_gramstain_nl("E. coli") # Negatieve staven


# Abbreviations known in the field
mo_genus("EHEC")           # Escherichia
mo_species("EHEC")         # coli
mo_subspecies("EHEC")      # EHEC
mo_fullname("EHEC")        # Escherichia coli (EHEC)

mo_genus("MRSA")           # Staphylococcus
mo_species("MRSA")         # aureus
mo_gramstain("MRSA")       # Positive cocci

mo_genus("VISA")           # Staphylococcus
mo_species("VISA")         # aureus


# Known subspecies
mo_genus("doylei")         # Campylobacter
mo_species("doylei")       # jejuni
mo_fullname("doylei")      # Campylobacter jejuni (doylei)


# Anaerobic bacteria
mo_genus("B. fragilis")    # Bacteroides
mo_species("B. fragilis")  # fragilis
mo_aerobic("B. fragilis")  # FALSE
}
\seealso{
\code{\link{microorganisms}}
}
