% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUCAST.R
\name{mo_property}
\alias{mo_property}
\title{Poperties of a microorganism}
\usage{
mo_property(bactid, property = "fullname")
}
\arguments{
\item{bactid}{ID of a microorganisme, like \code{"STAAUR} and \code{"ESCCOL}}

\item{property}{One of the values \code{bactid}, \code{bactsys}, \code{family}, \code{genus}, \code{species}, \code{subspecies}, \code{fullname}, \code{type}, \code{gramstain}, \code{aerobic}}
}
\description{
Poperties of a microorganism
}
\seealso{
\code{\link{microorganisms}}
}
