% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.symbol.R
\name{p.symbol}
\alias{p.symbol}
\title{Symbol of a p value}
\usage{
p.symbol(p, emptychar = " ")
}
\arguments{
\item{p}{p value}

\item{emptychar}{text to show when \code{p > 0.1}}
}
\value{
Text
}
\description{
Return the symbol related to the p value: 0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1. Values above \code{p = 1} will return \code{NA}.
}
\section{Read more on our website!}{

\if{html}{\figure{logo.png}{options: height=40px style=margin-bottom:5px} \cr}
On our website \url{https://msberends.gitlab.io/AMR} you can find \href{https://msberends.gitlab.io/AMR/articles/AMR.html}{a omprehensive tutorial} about how to conduct AMR analysis and find \href{https://msberends.gitlab.io/AMR/reference}{the complete documentation of all functions}, which reads a lot easier than in R.
}

