% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portion.R
\name{rsi}
\alias{rsi}
\title{Calculate resistance of isolates}
\usage{
rsi(ab1, ab2 = NULL, interpretation = "IR", minimum = 30,
  as_percent = FALSE, ...)
}
\arguments{
\item{ab1, ab2}{vector (or column) with antibiotic interpretations. It will be transformed internally with \code{\link{as.rsi}} if needed.}

\item{interpretation}{antimicrobial interpretation to check for}

\item{minimum}{the minimum allowed number of available (tested) isolates. Any isolate count lower than \code{minimum} will return \code{NA} with a warning. The default number of \code{30} isolates is advised by the Clinical and Laboratory Standards Institute (CLSI) as best practice, see Source.}

\item{as_percent}{a logical to indicate whether the output must be returned as a hundred fold with \% sign (a character). A value of \code{0.123456} will then be returned as \code{"12.3\%"}.}

\item{...}{deprecated parameters to support usage on older versions}
}
\description{
This function is deprecated. Use the \code{\link{portion}} functions instead.
}
